CreateVolume: use suggested name if possible
So far we had the unfortunate situation that the suggested name (e.g.,
as generated by a k8s PVC) did not match the underlying LINSTOR
resource name.

This was mainly due to a misreading of the spec, but also required some
infrastructure. With that in place, we can now generate names that are
equivalent, similar, or - in the "worst case" - a prefixed UUID.

This change does not break existing Volumes that got created with the
old naming.