manually create resource/volume definitions
Snapshots could not be restored since using ResourceGroups (RG). This was
because using the /spawn endpoint on RGs also creates a
VolumeDefinition (VD). On restoring the VD from the snapshot, linstor will
refuse, based on the fact that a VD with matching number already exists.

The fix is to manually create VDs instead of relying on /spawn.
The new volume creation process is:
* Create RG from volume params (== storage class params)
  There is only one RG for a storage class (as long as the RG name is set)
* Create ResourceDefinition for the new volume (and associate with RG)
* Create VolumeDefinition associated to ResourceDefinition

To create a new volume from a snapshot:
* Create RG from volume params (if not already present)
  There is only one RG for a storage class (as long as the RG name is set)
* Create ResourceDefinition for the new volume (and associate with RG)
* Restore VD from snapshot into ResourceDefinition
* Restore volumes based on VD and snapshot