add "Aux/" prefix to RG props
This fixes a convenience issue that bit some users already. LINSTOR aux
properties need the "Aux/" prefix on their keys. With that in place
users can specify it in "LINSTOR notation" with the prefix, or leave it
out and then the driver adds them. The LINSTOR command line client also
automatically adds the prefix under certain conditions, so it should be
fine to add this magic here as well.

The prefix is added to "replicasOnSame" and "replicasOnDifferent".

This allows:
parameters:
  autoPlace: "2"
  replicasOnDifferent: Aux/zone
  resourceGroup: linstor-volume-r2

as well as:
parameters:
  autoPlace: "2"
  replicasOnDifferent: zone
  resourceGroup: linstor-volume-r2