Reduce created topology keys for volumes
the topology feature is essential when working with volumes that are not
remote attachable (for example, without a drbd layer). A node publishes
a set of labels that can later be referenced by volumes to bind to a
specific node set.

A created volume can be accessible via 1-2 "topologies":
* node names: this is used to directly relate the backing disks created
  in LINSTOR to a set of nodes. This works for all LINSTOR managed resources.
* storage pool names: this is used in case remote access to a volume is
  possible. This depends on the wheter the storage class supports diskless
  volumes.

Using storage pool names greatly decreases the size of the generated PV spec,
since only a few nodes are directly referenced (those with backing disk
resources). In the old appoach, only using node names meant that any "diskless"
supporting volume would:
* have a list of _all_ nodes in the cluster in its spec
* would not be schedulable on nodes that joined the cluster after the volume
  was created (the list would not be updated at that point).