Update dev deployment to be up-to-date with 1.17
The last dev deployment example was not updated to support newer
kubernetes versions. Newer kubernetes distributions removed support
for some of the API versions which were still in use. This commit
updates the dev deployment to be in line with the 1.17 deployment.

Also adds a clarification that this project only deploys the CSI
components and links to the operator for more information.