remove losetup dependency for ROX support
In d2f23d6 we introduced ROX volume support, which works around DRBD's
autopromote feature by introducing a readonly loop-device so all access
is done readonly.

The code to handle detaching needed to clean up the loop device, which
relied on a specifc implementation of "losetup --list", which is not
compatible with older version (such as the one found in CentOS 7).
This meant that _all_ PVC failed to detach.

This commit switches to using "blockdev --setro" to ensure readonly access.
That in turn requires DRBD v9.0.26, which is not released yet. This should
not be a problem, as the feature was probably never used anyways.