Rework snapshots to work without metadata in RDs
The changes in detail:

Instead of storing snapshot information in resource definitions, use native
LINSTOR snapshots. This makes it easier to retry a timed out snapshot attempt.
A basic test was added using "old" format LINSTOR responses to ensure no
information is lost in transition.

To ensure stable snapshot ID between tries, the suggested name is either
used directly or a generated UUIDv5 based on the name is used. Since the CSI
snapshotter image will always suggest a compatible name, this is mostly done
for complying with the CSI spec.

In volume-from-volume scenario, re-use the snap-create and vol-from-snap logic
instead of implementing it in the LINSTOR client. This is done by first creating
a stable snapshot name for the new PVC, restoring using the existing
vol-from-snap logic and deleting the snapshot again.

The snapshots itself are only considered ready if LINSTOR reports the SUCCESS
flag. This prevents a lot of churn with snapshots that are not immediatly ready
being deleted over and over again.

Pagination is handled directly by the client, as we use the /view/snapshots
API directly.