Make snapshots independent from their source PV.
This commit enables deletion of PV while taken snapshots are kept around.
To implement this a couple of changes needed to be made:

Snapshots in LINSTOR are attached to a ResourceDefinition (RD), not to the
Resources directly. This means we can remove a PV by deleting the Resources
in LINSTOR without deleting the RD.

In turn, we cannot rely on the "volume.Info" struct to exist when
restoring from a snapshot. While it was queried, it actually did not serve
any purpose, the query is just removed.

Not deleting the RD on PV deletion requires another set of changes:
* Once the last RD "user" (Resource or Snapshot) is deleted, we need to
  delete the RD (and maybe also the Resource Group)
* We use aux props on the RD to persist important metadata. Specifically,
  it was also used to identify RDs and their matching PV. This metadata
  is now stripped on volume deletion.