only modify RG if parameters changed
Instead of essentially re-creating the RG on every reconcile call, we check for
changed values and only update the parts that diverged.

This change was made necessary by LINSTOR's autoplacer being triggered when
"updating" a RG used by a ResourceDefinition that had no resource left. Such
situation come up when deleting a PV with snapshot. The snapshot, and with it
the RD are left in LINSTOR, while all the resources are deleted.

Later, when trying to restore from the snapshot, a RG modify was triggered
which "updated" the placecount in LINSTOR (actually kept them the same, but
LINSTOR does not care). This triggered LINSTOR to place the resources again.

Even later, when the snapshot was removed, the RD was still left in LINSTOR as
the CSI driver sees the re-created resources and refuses to delete the RD,
thinking the resouces are part of an PV