only delete resources on Detach if created in Attach
We want to delete resources that were only created to facilitate an Attach()
request. The way this was determined was by checking if the given resource is
diskless or not. This presents a problem, as there are a number of ways
diskless resources get created:
* Linstor creates a TieBreaker resource
* The user specified that a Diskless resource should be placed on that node
* Attach() was called for a node were the resource was not present

From the point of view of Detach(), all these above cases look the same, and
so the resources were deleted even if they did not originate from the Attach.

This commit adds a new property to resources generated by Attach. This way
Detach can later check for the presence of this property, and only delete
the resource if:
* The resource has the "Attach" property
* The resource is still diskless (there could be cases in which LINSTOR
  migrated some replicas from a evicted node) Since we don't want to mess
  with the replication factor, diskfull resources are off-limits.