Remove failed attempts at creating a snapshot
In some cases snapshots are created but fail to become "successful",
for example when one of the satellite goes offline at the wrong moment.
In such cases, we want to delete the failed attempt, as otherwise
the CSI snapshotter will forever wait for the snapshot to get ready.
This then produces a lot (read: gigabytes) of logs, as it tries polling
the snapshot state again multiple times per second.

This commit adds logic to recognize such failed snapshots and remove+retry
them. Note that currently there is no way to tell the snapshot controller that
a snapshot can't ever succeed, for example when a storage driver without
snapshot support is used.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>