Try to ensure quorum is achieved when attaching
In certain circumstances it is necessary to create a diskfull resource to make
it usable. The bug report that introduced this variable is a good example:
* Our cluster has 4 identical nodes called A, B, C, D.
* A resource was placed in a typical 2 + 1 Tiebreaker configuration, let's say
  the diskfull resources are on A and B, the Tiebreaker on C.
* A Pod attaches on node A.
* Node A goes down/becomes unreachable/etc. That leaves us with 1 diskfull
  resource on B + 1 diskless on C.
* The Pod is deleted and a replacement scheduled on node D (i.e. after deletion
  by the HA Controller).
* Now, this method will be called, since we need to Attach on node D, and there
  is currently no resource on that node.
* Using a diskless resource will not work, as the newly creates resource would
  not get quorum. This is because the new diskless will only see 1 available
  and one unavailable replica, so it can't decide which is in the majority.
* Using a diskfull resource will work and sync up to the remaining diskfull
  resource on B, as in that case the new resources itself already counts
  for the quorum majority.

For this change we need to detect these situations. In general, the promotion
score reported by LINSTOR is a good indicator of "unavailable" vs "available"
replicas, even if the satellite itself is reachable.

This commit also adds some tests for the above procedure and other edge cases
on calling Attach(). The LINSTOR API is mocked with responses taken from a real
LINSTOR instance exhibiting these edge cases.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>