autoplacetopology: deal with incomplete topologies
When a user is using "AllowedTopologies", we might get incomplete topology
information (i.e. we only get key-values that are available on multiple nodes).
AutoPlaceTopology expected to always get the node name as one of the segments,
which meant that in the above case we would automatically fall back to a
normal autoplace.

This was not correct, as the user could for example specify:
"topology.kubernetes.io/zone=zone-1" as the only allowed topology,
which would then get ignored by autoplace.

The fix is to first try to extract the node name first and, if necessary, fall
back to directly querying the LINSTOR API for matching nodes.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>