capacity: track capacity for aggregate topology
if a storage class requests an aggregate topology (i.e. multiple nodes), we
should add the available storage capacity. This is basically the same fix
as applied in fa05b632a265009f3d5f481a9165c214ad9aabf4 to the volume scheduler.

Instead of expecting the topology segment to apply to a specific node, we try
to collect all the nodes that match the given segments.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>