support for creating backups to S3 via snapshot mechanism
In addition to the normal "in-cluster" snapshots, the user can also select
backups to S3 by providing parameters when creating the snapshot.

The parameters can be used to set up a LINSTOR "remote" object, which
represents a S3 endpoint. Any snapshot using those parameters will be
created using LINSTORs backup feature.

Restoring from a snapshot that only exists on a remote location does works,
as long as the "remote" is configured beforehand. The CSI spec does not
pass any parameters when registring an existing snapshot.

This feature requires LINSTOR 1.18.0 to work. Previous versions did not
allow for customizing the snapshot name. This made it impossible to create
an implementation that adheres to the CSI spec, as there was no way to
map the requested snapshot name to any form of stable snapshot ID.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>