store remote access policy in volume context
Some external components need to check the remote access policy. The policy
is only available in the parameters, which is derived from the storage class
at provision time. The storage class might be removed or changed at a later
time, making it impossible for these external components to derive the "real"
value to use.

Storing the value in the volume context means that is always available while
the volume is available, tying the remote access policy directly to the volume.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>