snapshots: only snapshots if nodes are online
snapshots with offline nodes create errors in LINSTOR. Since there is no
good back-off strategy employed by who-ever creates the snapshot requests,
we create linstor errors in a tight loop. To mitigate this issue, we check
the resource state before snapshotting. If LINSTOR does not report any
state, we can assume the node to be offline.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>