add option to change property -> topology mapping
In the future we might want to save additional information on node
properties that is not suitable to be a topology segment. For example,
it might not fit as a k8s label.

To allow this, we make the property mapping configurable, i.e. instead of
assuming all auxiliary properties should be applied as node labels, we
make the exact property prefix configurable.

With this change, the deployer (i.e. operator) can start to confine topology
information to a specific namespace, while having the option to store extra
information in regular auxiliary properties.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>