use golinstor for client initialization
golinstor already supports client configuration via environment variables.
For unknown reason, linstor-csi had it's own version of that initialization.

This commit removes the custom init code for the LINSTOR client for the most
part. We need to keep some routines to support the existing flags, as those
might be set by some users.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>