support deletion of snapshots after restore from backup
When a user requests a restore from an off-site backup, they might want the
local snapshot to be deleted after the restore.

We already have this option when creating the snapshot by using a special
parameter on the snapshot class. However, we do not have that information
available directly.

To support this, we need to lookup the snapshot content and snapshot class
ourselfes, then parse the parameters. This is all optional, as LINSTOR CSI
still needs to work outside kubernetes, or inside Kubernetes with incomplete
information.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>