docker: switch base to debian 12
Some tools that interact with the kernel are not happy when using old tools on
a new kernel. For example, turns out that resize2fs (the ext* resize tool)
refuses to resize a mounted FS on a new kernel because the FS there reports
some features it does not know about.

To fix, we use the latest debian stable release as our base image. Tools there
should be new enough for the future.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>