implement L2L backup shipping
Fill in the missing implementation for backup shipping to another LINSTOR
cluster. Because we can't "wish" for a specific snapshot name, we store a
mapping of requested snapshot name to actually taken snapshot name in
a special LINSTOR KV.

Other than that, it is simply a matter of wiring up the necessary plumbing.

This currently cannot restore a backup once it is shipped and the local
snapshot deleted, and it is unlikely that such a thing could be implemented,
as LINSTOR does not support downloading a snapshot once shipped.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>