really disable fallback mounting to RO
Passing the "rw" option is not enough. Careful reading of the manpage, verified
by tests and reading of the source code of util-linux reveals that disabling
the RO fallback only happens when you pass the "-w" flag.

Then, also fix "ro" mounts in case of XFS, as XFS attempts to write even if
passed the "ro" flag without the "norecovery" option.

Fixes: 6deb565 ("disable fallback to RO mounts for RW volumes")
Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>