skip LINSTOR KV for local and S3 snapshots
For L2L snapshots, we need to store the name of the created snapshot, as
we can't influence the LINSTOR generated snapshot name for those. Storing the
name means we can find the created snapshot again should the request need to
be retried because of a timeout.

For local and S3 snapshots, we can just pass the snapshot ID from CSI to
LINSTOR, meaning we don't need to external name mapping. So we can also
skip the cleanup step.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>