remove dependency on "github.com/haySwim/data"
From Kubernetes side, everything is bytes. LINSTOR expects KiB. While using
proper types to protect the units is nice in theory, the current implementation
by the "data" module we used was less nice.

There where a lot of arcane casts from one type to another, when all we need
is some basic * 1024 or / 1024.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>