# This is a combination of 2 commits. # This is the 1st commit message:
Add support for RWX volumes via NFS

The NFS export is handled by a separate component that is configured via a
config map. When a new RWX volume is created, the config map is updated with
the necessary information on which DRBD resource is to be used, as well as
configuring the Service that is used to provide the service IP address of
the NFS export.

The main changes are:
* Ensuring that ControllerPublishVolume and NodePublishVolume calls to not
  try to interact with node-local devices in case of NFS. Instead, the NFS
  export is expected to be available when call NodePublishVolume.
* Ensuring that LINSTOR CSI can handle resources with multiple volumes. This
  is needed as the NFS exports have a second volume used for client recovery.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>

# Conflicts:
#	pkg/client/linstor.go

# Conflicts:
#	pkg/driver/driver.go

# The commit message #2 will be skipped:

# fixup! Add support for RWX volumes via NFS