sanity check that block volumes are usable
We want to ensure a mapped block volume are usable. Since we only bind-mount
the device to a new location, we never verify that the device is usable. This
may lead to a device that has no quorum being used. Depending on the workload
using that device, this may go unnoticed.

So we do a simple sanity check by opening the device. A "rw" device opening
fails immediately on open, a "ro" device might only fail on reading a specific
sector. Since we can't really read the whole device, no additional checks are
added for this case.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>