nfs: add start-stop-reactor command
We use DRBD Reactor to control our NFS servers. However, in the case the node
is drained for maintenance, we want to stop the NFS servers. Since the way
we deploy the DRBD Reactor container means we do not get stopped by Kubernetes,
we have to manually stop the DRBD Reactor.

This is done by the start-stop-reactor command and service: if the node we are
running on is schedulable, we start the DRBD Reactor so that it can run NFS
servers. If the Node gets unschedulable, we stop the DRBD Reactor, also
stopping all NFS servers.

As a second optimization: while we could wait for Kubernetes to propagate
updates to the DRBD Reactor config manually, this can take a long time based
on different Kubernetes settings. So the start-stop-reactor job also watches
for changes in the ConfigMap storing the DRBD Reactor config snippets.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>